ekart-aws-localstack/lambda-functions/cart-api/requirements.txt
boto3>=1.26.0
PyJWT>=2.10.1
